thumbnail: Move get_image_thumbnail_path and split_thumbnail_path.
